Windows will now compile and work right(for MSVC anyways)

+++ classes/containers/inventory.hxx
@@ -61,7 +61,7 @@ class Inventory {
 			}
 		}
 		//set the size to the total number.
-		this->num_items = items.size();
+		this->num_items = static_cast<unsigned short>(items.size());
 	}
 
 	/**
